move deprecated package to internal directory
