+++ scripts/packages.sh
@@ -3,4 +3,4 @@ set -eu
 
 source scripts/_prepare.sh
 
-echo "$(${LOCAL_GO} list ./... | \grep -v 'vendor' | \grep -v '/examples' | \grep -v '/deprecated')"
+echo "$(${LOCAL_GO} list ./... | \grep -v 'vendor' | \grep -v '/examples')"
